complaint-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
python-jose[cryptography]